Design Customized

+++ src/Pages/Dashboard/AllUsers/AllUsers.js
@@ -1,46 +1,48 @@
 import { useQuery } from '@tanstack/react-query';
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import toast from 'react-hot-toast';
 import ConfirmationModal from '../../Shared/ConfirmationModal/ConfirmationModal';
 import Loading from '../../Shared/Loading/Loading';
 
 const AllUsers = () => {
 
-    // const [users, setUsers] = useState([]);
-    const [deletingUser, setDeletingUser] = useState(null);
+    const [users, setUsers] = useState([]);
+    // const [deletingUser, setDeletingUser] = useState(null);
 
-    const closeModal = () => {
-        setDeletingUser(null);
-    }
+    // const closeModal = () => {
+    //     setDeletingUser(null);
+    // }
 
-    const { data: users, isLoading, refetch } = useQuery({
-        queryKey: ['users'],
-        queryFn: async () => {
-            try {
-                const res = await fetch('https://laptop-zone-server.vercel.app/users', {
-                    headers: {
-                        authorization: `bearer ${localStorage.getItem('accessToken')}`
-                    }
-                });
-                const data = await res.json();
-                return data;
-            }
-            catch (error) {
+    // const { data: users, isLoading, refetch } = useQuery({
+    //     queryKey: ['users'],
+    //     queryFn: async () => {
+    //         try {
+    //             const res = await fetch('http://localhost:5000/allusers', {
+    //                 headers: {
+    //                     authorization: `bearer ${localStorage.getItem('accessToken')}`
+    //                 }
+    //             });
+    //             const data = await res.json();
+    //             return data;
+    //         }
+    //         catch (error) {
 
-            }
-        }
-    });
+    //         }
+    //     }
+    // });
 
-    console.log(users);
+    // console.log(users);
 
-    // useEffect(() => {
-    //     fetch('https://laptop-zone-server.vercel.app/users')
-    //         .then(res => res.json())
-    //         .then(data => setUsers(data))
-    // }, []);
+    useEffect(() => {
+        fetch('http://localhost:5000/allusers')
+            .then(res => res.json())
+            .then(data => setUsers(data))
+    }, []);
+
+    const {_id:id} = users;
 
     // const handleMakeAdmin = id => {
-    //     fetch(`https://laptop-zone-server.vercel.app/users/admin/${id}`, {
+    //     fetch(`http://localhost:5000/users/admin/${id}`, {
     //         method: 'PUT',
     //         headers: {
     //             authorization: `bearer ${localStorage.getItem('accessToken')}`
@@ -55,44 +57,46 @@ const AllUsers = () => {
     //         })
     // };
 
-    // const handleDelete = id => {
-    //     const proceed = window.confirm("Are you sure you want to delete the User?")
-    //     if (proceed) {
-    //         fetch(`https://laptop-zone-server.vercel.app/users/${id}`,
-    //             {
-    //                 method: 'DELETE'
-    //             })
-    //             .then(res => res.json())
-    //             .then(data => {
-    //                 console.log(data);
-    //                 if (data.deletedCount > 0) {
-    //                     toast.success('deleted successfully');
-    //                     const remaining = users.filter(rev => rev._id !== id);
-    //                     setUsers(remaining);
-    //                 }
-    //             })
-    //     }
+    const handleDelete = id => {
+        const proceed = window.confirm("Are you sure you want to delete the User?")
+        console.log(id);
+        if (proceed) {
+            fetch(`http://localhost:5000/users/${users.id}`,
+                {
+                    method: 'DELETE'
+                })
+                .then(res => res.json())
+                .then(data => {
+                    console.log(data);
+                    if (data.deletedCount > 0) {
+                        toast.success('deleted successfully');
+                        const remaining = users.filter(rev => rev._id !== id);
+                        setUsers(remaining);
+                    }
+                })
+        }
+    }
+
+    // const handleDelete = user => {
+
+    //     fetch(`http://localhost:5000/allusers/${user._id}`, {
+    //         method: 'DELETE',
+    //         headers: {
+    //             authorization: `bearer ${localStorage.getItem('accessToken')}`
+    //         }
+    //     })
+    //         .then(res => res.json())
+    //         .then(data => {
+    //             if (data.deletedCount > 0) {
+    //                 refetch();
+    //                 toast.success('User deleted successfully')
+    //             }
+    //         })
     // }
 
-    const handleDelete = user => {
-        fetch(`https://laptop-zone-server.vercel.app/users/${user._id}`, {
-            method: 'DELETE',
-            headers: {
-                authorization: `bearer ${localStorage.getItem('accessToken')}`
-            }
-        })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0) {
-                    refetch();
-                    toast.success('User deleted successfully')
-                }
-            })
-    }
-
-    if (isLoading) {
-        return <Loading></Loading>
-    }
+    // if (isLoading) {
+    //     return <Loading></Loading>
+    // }
 
     return (
         <div>
@@ -122,7 +126,7 @@ const AllUsers = () => {
                     </tbody>
                 </table>
             </div>
-            {
+            {/* {
                 deletingUser && <ConfirmationModal
                     title={`Are you sure you want to delete?`}
                     message={`If you delete. It cannot be undone.`}
@@ -132,7 +136,7 @@ const AllUsers = () => {
                     closeModal={closeModal}
                 >
                 </ConfirmationModal>
-            }
+            } */}
         </div>
     );
 };
